+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ### 2.0.3
 + [x] Server: added support for time interval string values for parameter `restartAfterNotSynced`  
 + [x] Server: improved check for Fork. Now check work if all control values gt zero
-+ [x] Server: default config value for `blockDiffToRestart` changed to `10`
++ [x] Server: default value for config parameter `blockDiffToRestart` changed to `10`
++ [x] Server: default value for config parameter `blockDiff` changed to `3`
 
 ### 2.0.2
 + [x] Server: added restart by long uptime with config parameter `restartAfterUptime`. The value must be defined in milliseconds or internal time string format . Default value is `0` (no restart)